bakery-10 [beta]: implemented goods prices table for the Clients.

+++ retrade/src/main/java/com/tverts/retrade/data/TradeStoreModelData.java
@@ -56,7 +56,7 @@ public class TradeStoreModelData implements ModelData
 	}
 
 	@XmlElement
-	public long                getGoodsNumber()
+	public int                 getGoodsNumber()
 	{
 		return bean(GetGoods.class).
 		  countGoodUnits(getModel());
